Added display_snippets.h, a solit base of basic display related operations, and build display.h newly on it.
Display_snippets.h provides a base that is used by most of the macros in display.h . This allows to quickly change basic implemantation for ever higher abstacted funktion. Also readability in display.h increased a lot. The granulanity of display_sippets.h is optimized to allow the most effitient implementation for every higher abstacted macro in display.h .